+++ plugin/emacs/flymake-quicklintjs.el
@@ -39,6 +39,11 @@
   :group 'flymake-quicklintjs
   :type '(repeat string))
 
+(defcustom flymake-quicklintjs-experimental-typescript nil
+  "Enable experimental TypeScript support (EXPERIMENTAL; subject to change in future versions of quick-lint-js)."
+  :group 'flymake-quicklintjs
+  :type '(boolean))
+
 (defvar-local flymake-quicklintjs--proc nil
   "Internal variable for `flymake-quicklintjs'")
 
@@ -74,8 +79,11 @@ REPORT-FN is Flymake's callback."
            :command `(,flymake-quicklintjs-program
                       ,@(let ((file (buffer-file-name)))
                           (if file
-                            `("--path-for-config-search" ,file)
+                            `("--stdin-path" ,file)
                             ()))
+                      ,@(if flymake-quicklintjs-experimental-typescript
+                            `("--language=experimental-default")
+                          ())
                       "--stdin" "--output-format=emacs-lisp"
                       ,@flymake-quicklintjs-args)
            :sentinel